LF-5114 Extract the offline indicator to its own component
